Salarios y recargos

+++ src/main/java/co/com/johan/green/dto/SalariosRecargosDTO.java
@@ -1,25 +1,22 @@
 package co.com.johan.green.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import java.util.List;
-import java.util.ArrayList;
 
 /**
-  * @generated
-  *  @author Johan Lopez
-  */
+ * @generated @author Johan Lopez
+ */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class SalariosRecargosDTO {
 
-    public SalariosRecargosDTO(){
+    public SalariosRecargosDTO() {
         //constructor base
     }
-    
-    public SalariosRecargosDTO(Long id){
-        this.id=id;
+
+    public SalariosRecargosDTO(Long id) {
+        this.id = id;
     }
 
-	private Long id;
+    private Long id;
 
     public Long getId() {
         return id;
@@ -28,149 +25,138 @@ public class SalariosRecargosDTO {
     public void setId(Long id) {
         this.id = id;
     }
-	
-    
+
     /**
-    * @generated
-    */
+     * @generated
+     */
     private String descripcion;
-    
+
     /**
-    * @generated
-    */
+     * @generated
+     */
     private String unidad;
-    
+
     /**
-    * @generated
-    */
-    private Integer cantidad;
-    
+     * @generated
+     */
+    private Double cantidad;
+
     /**
-    * @generated
-    */
+     * @generated
+     */
     private Double valor;
-    
+
     /**
-    * @generated
-    */
+     * @generated
+     */
     private Double total;
-    
+
     /**
-    * @generated
-    */
+     * @generated
+     */
     private Boolean activo;
-    
-    
+
     /**
-    * @generated
-    */
+     * @generated
+     */
     private CargoDTO cargo;
-    
+
     /**
-    * @return
-    * @generated
-    */
+     * @return @generated
+     */
     public String getDescripcion() {
         return this.descripcion;
     }
-    
+
     /**
-    * @param
-    * @generated
-    */
+     * @param @generated
+     */
     public void setDescripcion(String descripcion) {
         this.descripcion = descripcion;
     }
+
     /**
-    * @return
-    * @generated
-    */
+     * @return @generated
+     */
     public String getUnidad() {
         return this.unidad;
     }
-    
+
     /**
-    * @param
-    * @generated
-    */
+     * @param @generated
+     */
     public void setUnidad(String unidad) {
         this.unidad = unidad;
     }
+
     /**
-    * @return
-    * @generated
-    */
-    public Integer getCantidad() {
+     * @return @generated
+     */
+    public Double getCantidad() {
         return this.cantidad;
     }
-    
+
     /**
-    * @param
-    * @generated
-    */
-    public void setCantidad(Integer cantidad) {
+     * @param @generated
+     */
+    public void setCantidad(Double cantidad) {
         this.cantidad = cantidad;
     }
+
     /**
-    * @return
-    * @generated
-    */
+     * @return @generated
+     */
     public Double getValor() {
         return this.valor;
     }
-    
+
     /**
-    * @param
-    * @generated
-    */
+     * @param @generated
+     */
     public void setValor(Double valor) {
         this.valor = valor;
     }
+
     /**
-    * @return
-    * @generated
-    */
+     * @return @generated
+     */
     public Double getTotal() {
         return this.total;
     }
-    
+
     /**
-    * @param
-    * @generated
-    */
+     * @param @generated
+     */
     public void setTotal(Double total) {
         this.total = total;
     }
+
     /**
-    * @return
-    * @generated
-    */
+     * @return @generated
+     */
     public Boolean getActivo() {
         return this.activo;
     }
-    
+
     /**
-    * @param
-    * @generated
-    */
+     * @param @generated
+     */
     public void setActivo(Boolean activo) {
         this.activo = activo;
     }
-    
-	/**
-	* @return
-	* @generated
-	*/
-	public CargoDTO getCargo() {
-	    return this.cargo;
-	}
-	
-	/**
-	* @param
-	* @generated
-	*/
-	public void setCargo(CargoDTO cargo) {
-	    this.cargo = cargo;
-	}
-	
+
+    /**
+     * @return @generated
+     */
+    public CargoDTO getCargo() {
+        return this.cargo;
+    }
+
+    /**
+     * @param @generated
+     */
+    public void setCargo(CargoDTO cargo) {
+        this.cargo = cargo;
+    }
+
 }
